requirements.txt
pandas==1.2.0
streamlit==0.81.1
numpy==1.19.3
Pillow==8.2.0
PyDrive==1.3.1
streamlit_cropper==0.1.3
boto3==1.17.81
diskcache==5.2.1
